Thanks to Domenico Carbotta who sent Jeroen and I these nice python scripts in email.  This makes a really nice HTML output when you run a python script.  It also catches errors and provides clickable links on a backtrace.
Inspired by his script I added an additional hack that overrides the input and raw_input commands in the environment set up for the script.  The result is that if you call input or raw_input in your script you will now get a dialog box asking you for input.  The dialogs are pretty simple I just used EasyDialogs (again thanks to Domenico).  To get EasyDialogs to work you must run pythonw.
The HTML window reminds you of this if you have set your TM_PYTHON variable to something else.

I moved the cmd-r binding to this new version, but left the old command in place in case others prefer the old output.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Python.tmbundle@1464 dfb7d73b-c2ec-0310-8fea-fb051d288c6d