I've tried to clean up naming and placement of these Python commands, since it really was/is a mess :)
I also made minor changes to the commands, but the functionality should remain (just tried to simplify a few things).

I removed the old Python run command, we did that for Ruby as well. I also spelled it Unit Test in two words, trying to mimic the structure of the Ruby bundle, though this is currently inconsistent with the Python templates.

Apologies in advance if I broke something.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Python.tmbundle@1980 dfb7d73b-c2ec-0310-8fea-fb051d288c6d