• Fix handling of script's with hashbangs.
If the script contains a hashbang, we use that to form the basis of the command to launch the script (but still append any additional args). To do this, we run Shellwords over the hashbang instead of calling `executable`. The python and ruby scriptmates have been updated to be compatible with this change.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Python.tmbundle@9642 dfb7d73b-c2ec-0310-8fea-fb051d288c6d