Run command: Output links even for untitled documents
I assume the reason links were skipped for “TM_FILE_IS_UNTITLED” was because the original author didn’t bother adding the code to remove the URL-part of the txmt://-links. Since TextMate.executor handles conversion of the links, there should be no need to special-case the “untitled document” case in this file.