Update “Execute Line” command
Incase of errors from ‘eval’ we simply show a tool tip with what we tried to evaluate rather than try to dump the stack. The stack dump code wasn’t working, closes #15.

When evaluating the selection we now insert ‘ #=> ’ before the result. There was a typo in the code so previously a newline was always inserted before the result.