Trigger “indent current line” after inserting a colon
Since we have disabled indent-as-you-type for Python, we don’t get the automatic decrease of indent when typing the else keyword in an indented if block.

This commit adds the colon as a special case for triggering automatic indent of the current line. Ideally there would be an option to only partially disable TextMate’s indent-as-you-type so that this macro was not required.