Altered folding markers to match {}, (), []
TextMate 1.x folded code, even Python code, in {}, (), and []. This is pretty handy for long lists, constructor use, etc. The current default start/stop markers only fold on triple-quoted docstrings. I kept those patterns, but added to each one so they'll now also match on {}, (), and []. Supports lines with comments to the right of them, and closing markers followed by commas in case it's being declared as part of a longer list or tuple.